+++ changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
-Last change: 10-JUN-2026 CEST
+Last change: 11-JUN-2026 CEST
 
 ## 0.10.2
+- Rotations use quaterions for unambiguous workflow, fixing bugs [#673](https://github.com/jungmannlab/picasso/issues/673), [#674](https://github.com/jungmannlab/picasso/issues/674) and [#675](https://github.com/jungmannlab/picasso/issues/675)
 - Fixed "Best fitting combination" button in SPINNA ([#676](https://github.com/jungmannlab/picasso/issues/676))
 - Fixed 3D calibration when frame range is user-specified
 - Fixed redoing 3D calibration when identification parameters change